Consolidate CI detection into core platform helper
Move CIEnvironmentDetectorForTelemetry from the Telemetry extension into

Microsoft.Testing.Platform as CIEnvironmentDetector (instance-based with

IEnvironment for testability, marked [Embedded] so the Telemetry extension

source-links its own copy instead of relying on InternalsVisibleTo).

Use the consolidated detector in TerminalOutputDevice instead of the inline

TF_BUILD || GITHUB_ACTIONS check, closing the long-standing TODO that asked

to reuse the telemetry CI detection (microsoft/testfx#5533). The terminal

now switches to its CI-friendly SimpleAnsi mode for the broader set of CI

providers detected by the standard .NET CI-detection list. Use --no-ansi

if you prefer plain output.

Also reword SimpleAnsiTerminal documentation to be CI-generic rather than

AzDO-specific.
